% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coretta2018itaegg.R
\docType{package}
\name{coretta2018itaegg}
\alias{coretta2018itaegg}
\title{coretta2018itaegg: Data Package of Vowel duration, voicing duration, and
vowel height: Acoustic and articulatory data from Italian (Coretta 2018)}
\description{
This package provides datasets containing acoustic measurements of Italian.
The scope of the data is to investigate the relation between vowel height,
vowel duration, and voicing duration.
}
