% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{formants}
\alias{formants}
\title{Data on vowel formants and duration in Italian.}
\format{
A tibble with 3053 observations and 38 variables.
}
\usage{
formants
}
\description{
A dataset with duration and formants measures from 19 speakers of Italian.
}
\keyword{datasets}
